cleanup from Gemini

+++ libraries/Woyak/Color.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <string>
+#include <Arduino.h>
 
 #if defined(_Adafruit_GRAYOLED_H_)
 
@@ -46,6 +46,7 @@ enum class Color : uint16_t;
 
 namespace Color565
 {
+   ///
    /// <summary>
    /// Converts RGB color components to 565-bit color format.
    /// </summary>
@@ -53,6 +54,7 @@ namespace Color565
    /// <param name="green">Green component (0-255)</param>
    /// <param name="blue">Blue component (0-255)</param>
    /// <returns>Color in 565 format</returns>
+   ///
    constexpr Color fromRGB(uint8_t red, uint8_t green, uint8_t blue)
    {
       return (Color) (((red & 0xF8) << 8) | ((green & 0xFC) << 3) | (blue >> 3));
@@ -95,46 +97,78 @@ enum class Color : uint16_t
 
 namespace Color565
 {
-   /// <summary>Extracts red component (0-255) from 565-bit color.</summary>
+   ///
+   /// <summary>
+   /// Extracts red component (0-255) from 565-bit color.
+   /// </summary>
    /// <param name="color">Color value (16-bit)</param>
    /// <returns>Red component (0-255)</returns>
+   ///
    uint8_t getR(uint16_t color)
    {
       return (uint8_t) (255 * ((color >> 11) & 0x1F) / 31.0);
    }
 
-   /// <summary>Extracts red component (0-255) from Color.</summary>
+   ///
+   /// <summary>
+   /// Extracts red component (0-255) from Color.
+   /// </summary>
+   /// <param name="color">Color value</param>
+   /// <returns>Red component (0-255)</returns>
+   ///
    uint8_t getR(Color color) { return getR((uint16_t)color); }
 
-   /// <summary>Extracts green component (0-255) from 565-bit color.</summary>
+   ///
+   /// <summary>
+   /// Extracts green component (0-255) from 565-bit color.
+   /// </summary>
    /// <param name="color">Color value (16-bit)</param>
    /// <returns>Green component (0-255)</returns>
+   ///
    uint8_t getG(uint16_t color)
    {
       return (uint8_t) (255 * ((color >> 5) & 0x3F) / 63.0);
    }
 
-   /// <summary>Extracts green component (0-255) from Color.</summary>
+   ///
+   /// <summary>
+   /// Extracts green component (0-255) from Color.
+   /// </summary>
+   /// <param name="color">Color value</param>
+   /// <returns>Green component (0-255)</returns>
+   ///
    uint8_t getG(Color color) { return getG((uint16_t)color); }
 
-   /// <summary>Extracts blue component (0-255) from 565-bit color.</summary>
+   ///
+   /// <summary>
+   /// Extracts blue component (0-255) from 565-bit color.
+   /// </summary>
    /// <param name="color">Color value (16-bit)</param>
    /// <returns>Blue component (0-255)</returns>
+   ///
    uint8_t getB(uint16_t color)
    {
       return (uint8_t) (255 * ((color & 0x1F) / 31.0));
    }
 
-   /// <summary>Extracts blue component (0-255) from Color.</summary>
+   ///
+   /// <summary>
+   /// Extracts blue component (0-255) from Color.
+   /// </summary>
+   /// <param name="color">Color value</param>
+   /// <returns>Blue component (0-255)</returns>
+   ///
    uint8_t getB(Color color) { return getB((uint16_t)color); }
 
+   ///
    /// <summary>
    /// Blends two 565-bit colors using linear interpolation.
    /// </summary>
    /// <param name="c1">First color</param>
    /// <param name="c2">Second color</param>
-   /// <param name="ratio">Blend ratio (0.0=c1, 1.0=c2)</param>
+   /// <param name="ratio">Blend ratio (0.0 = c1, 1.0 = c2)</param>
    /// <returns>Blended color</returns>
+   ///
    Color blend(uint16_t c1, uint16_t c2, float ratio)
    {
       uint8_t r = constrain(getR(c1) + ratio * ((int16_t)getR(c2) - getR(c1)), 0, 255);
@@ -143,21 +177,45 @@ namespace Color565
       return fromRGB(r, g, b);
    }
 
-   /// <summary>Blends two Color values using linear interpolation.</summary>
+   ///
+   /// <summary>
+   /// Blends two Color values using linear interpolation.
+   /// </summary>
    /// <param name="c1">First color</param>
    /// <param name="c2">Second color</param>
-   /// <param name="ratio">Blend ratio (0.0=c1, 1.0=c2)</param>
+   /// <param name="ratio">Blend ratio (0.0 = c1, 1.0 = c2)</param>
    /// <returns>Blended color</returns>
+   ///
    Color blend(Color c1, Color c2, float ratio) { return blend((uint16_t)c1, (uint16_t)c2, ratio); }
 
-   /// <summary>Blends color values using linear interpolation (overload variants).</summary>
-   Color blend(uint16_t c1, Color c2, float ratio) { return blend((uint16_t)c1, (uint16_t)c2, ratio); }
+   ///
+   /// <summary>
+   /// Blends color values using linear interpolation (overload variant).
+   /// </summary>
+   /// <param name="c1">First color (16-bit)</param>
+   /// <param name="c2">Second color</param>
+   /// <param name="ratio">Blend ratio (0.0 = c1, 1.0 = c2)</param>
+   /// <returns>Blended color</returns>
+   ///
+   Color blend(uint16_t c1, Color c2, float ratio) { return blend(c1, (uint16_t)c2, ratio); }
 
-   /// <summary>Blends color values using linear interpolation (overload variants).</summary>
-   Color blend(Color c1, uint16_t c2, float ratio) { return blend((uint16_t)c1, (uint16_t)c2, ratio); }
+   ///
+   /// <summary>
+   /// Blends color values using linear interpolation (overload variant).
+   /// </summary>
+   /// <param name="c1">First color</param>
+   /// <param name="c2">Second color (16-bit)</param>
+   /// <param name="ratio">Blend ratio (0.0 = c1, 1.0 = c2)</param>
+   /// <returns>Blended color</returns>
+   ///
+   Color blend(Color c1, uint16_t c2, float ratio) { return blend((uint16_t)c1, c2, ratio); }
 
-   /// <summary>Prints the RGB components of a Color to Serial.</summary>
+   ///
+   /// <summary>
+   /// Prints the RGB components of a Color to Serial.
+   /// </summary>
    /// <param name="color">Color to print</param>
+   ///
    void print(Color color)
    {
       Serial.print("R:");
@@ -168,20 +226,32 @@ namespace Color565
       Serial.print(getB(color));
    }
 
-   /// <summary>Prints the RGB components of a 565-bit color to Serial.</summary>
+   ///
+   /// <summary>
+   /// Prints the RGB components of a 565-bit color to Serial.
+   /// </summary>
    /// <param name="color">Color value (16-bit)</param>
+   ///
    void print(uint16_t color) { print((Color)color); }
 
-   /// <summary>Prints the RGB components of a Color to Serial, followed by a newline.</summary>
+   ///
+   /// <summary>
+   /// Prints the RGB components of a Color to Serial, followed by a newline.
+   /// </summary>
    /// <param name="color">Color to print</param>
+   ///
    void println(Color color)
    {
       print(color);
       Serial.println();
    }
 
-   /// <summary>Prints the RGB components of a 565-bit color to Serial, followed by a newline.</summary>
+   ///
+   /// <summary>
+   /// Prints the RGB components of a 565-bit color to Serial, followed by a newline.
+   /// </summary>
    /// <param name="color">Color value (16-bit)</param>
+   ///
    void println(uint16_t color) { println((Color)color); }
 }
